requirements.txt
backports.functools-lru-cache==1.6.4
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
httplib2==0.19.0
idna==2.8
kiwisolver==1.1.0
matplotlib==2.2.5
numpy==1.16.6
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.22.0
rsa==4.7
six==1.14.0
uritemplate==3.0.1
urllib3==1.26.5
